+++ lib/ooxml_parser/docx_parser/docx_data/document_structure/numbering/numbering_definition.rb
@@ -2,7 +2,7 @@ require_relative 'numbering_definition/abstract_numbering_id'
 module OoxmlParser
   # This element specifies a unique instance of numbering information that can be referenced by zero or more
   # paragraphs within the parent WordprocessingML document.
-  class NumberingDefinition
+  class NumberingDefinition < OOXMLDocumentObject
     # @return [Integer] num id
     attr_accessor :id
     # @return [AbstractNumberingId] abstract numbering id
@@ -11,23 +11,21 @@ module OoxmlParser
     # Parse NumberingDefinition data
     # @param [Nokogiri::XML:Element] node with NumberingDefinition data
     # @return [NumberingDefinition] value of Abstract Numbering data
-    def self.parse(node)
-      num = NumberingDefinition.new
-
+    def parse(node)
       node.attributes.each do |key, value|
         case key
         when 'numId'
-          num.id = value.value.to_f
+          @id = value.value.to_f
         end
       end
 
       node.xpath('*').each do |numbering_child_node|
         case numbering_child_node.name
         when 'abstractNumId'
-          num.abstract_numbering_id = AbstractNumberingId.parse(numbering_child_node)
+          @abstract_numbering_id = AbstractNumberingId.new(parent: self).parse(numbering_child_node)
         end
       end
-      num
+      self
     end
   end
 end